requirements.txt
streamlit==1.29.0
pytesseract==0.3.10
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
streamlit-paste-button==0.1.0
httpx==0.27.0
